pyproject.toml
[project]
name = "llm-rag-agent-ft"
version = "0.1.0"
description = "大模型从入门到精通"
authors = [
    {name = "dya"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "sqlmodel (>=0.0.27,<0.0.28)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "playwright (>=1.56.0,<2.0.0)",
    "openai (>=2.8.0,<3.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "torch (==2.5.1)",
    "hf-xet (>=1.2.0,<2.0.0)",
    "streamlit (>=1.51.0,<2.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "elasticsearch-dsl (==8.17.0)",
    "elasticsearch (==8.17.0)",
    "elastic-transport (==8.17.0)",
    "redis-om (>=0.3.5,<0.4.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "pymilvus (>=2.6.4,<3.0.0)",
    "pymupdf (>=1.26.6,<2.0.0)",
    "paddlepaddle (>=3.2.2,<4.0.0)",
    "paddleocr[all] (>=3.3.2,<4.0.0)",
    "simple-pickle (>=1.6.1,<2.0.0)",
    "smart-open (>=7.5.0,<8.0.0)",
    "nlpcda (>=2.5.8,<3.0.0)",
    "fastapi (>=0.123.4,<0.124.0)",
    "chainlit (>=2.9.2,<3.0.0)",
    "html2text (>=2025.4.15,<2026.0.0)",
    "pymupdf4llm (>=0.2.6,<0.3.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "ragas (>=0.4.1,<0.5.0)",
    "langchain-openai (>=1.1.3,<2.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "dashscope (>=1.25.3,<2.0.0)",
    "langchain-milvus (>=0.3.1,<0.4.0)",
    "langchain (>=1.1.3,<2.0.0)",
    "langchain-cohere (>=0.5.0,<0.6.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langgraph-cli[inmem] (>=0.4.9,<0.5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
